+++ data.h
@@ -7,12 +7,12 @@ protected:
 	std::string title;
 	std::string creator;
 	int year;
-	int duration; //minutes
+	double duration; //minutes
 
 	data(const std::string& title,
 		const std::string& creator,
 		int year, 
-		int duration);
+		double duration);
 
 public:
 	virtual ~data() = default;
@@ -20,7 +20,7 @@ public:
 	std::string get_title() const;
 	std::string get_creator() const;
 	int get_year() const;
-	int get_duration() const;
+	double get_duration() const;
 
 	virtual void print() const = 0;
 };
