+++ tvEpisode.cpp
@@ -4,7 +4,7 @@ tv_episode::tv_episode(const std::string& title,
 	const std::string& show_title,
 	const std::string& creator, int year, 
 	int season_num, int episode_num,
-	int duration, int rating) 
+	double duration, int rating)
 	: data(title, creator, year, duration) {\
 
 	if (show_title.length() > 0) {
